+++ src/Audit.NET/AuditEvent.cs
@@ -29,6 +29,13 @@ namespace Audit.Core
         [JsonIgnore(Condition = JsonIgnoreCondition.WhenWritingNull)]
 #endif
         public AuditEventEnvironment Environment { get; set; }
+        
+#if NET5_0_OR_GREATER
+        /// <summary>
+        /// The current distributed tracing activity information 
+        /// </summary>
+        public AuditActivityTrace Activity { get; set; }
+#endif
 
         /// <summary>
         /// The extension data.