+++ test/Audit.UnitTest/HttpClientTests.cs
@@ -29,8 +29,7 @@ namespace Audit.UnitTest
         [SetUp]
         public void Setup()
         {
-            Audit.Core.Configuration.AuditDisabled = false;
-            Audit.Core.Configuration.ResetCustomActions();
+            Audit.Core.Configuration.Reset();
         }
 
         private Action<IAuditClientHandlerConfigurator> config = _ => _